Rename 'test' package to 'tests'
