+++ decompiler/tests/token_extraction_test.go
@@ -1,4 +1,4 @@
-package test
+package tests
 
 import (
 	"github.com/byxor/NeverScript/decompiler/tokens"
